+++ youkai-server/src/main/java/com/github/kazuhitom/youkai/server/presentation/security/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.github.kazuhitom.youkai.server;
+package com.github.kazuhitom.youkai.server.presentation.security;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
